Refactored code to make it abit DRY
